Minor bug fixes.

+++ gbpLib/gbpFITS/gbpFITS.h
@@ -2,7 +2,7 @@
 #define GBPFITS_AWAKE
 #include <fitsio.h>
 
-int read_image_FITS(void **image,SID_Datatype *dtype,int *n_D,long **D,char *filename);
+int read_image_FITS(void **image,SID_Datatype *dtype,int *n_D,int **D,char *filename);
 int write_image_FITS(void *image,SID_Datatype dtype,int n_D,long *D,char *filename);
 
 #endif
